+++ includes/auth.php
@@ -40,6 +40,7 @@ function auth_login(array $user): void
     auth_start_session();
     session_regenerate_id(true);
     $_SESSION['user_id'] = $user['id'];
+    user_record_login($user['id']);
 }
 
 function auth_logout(): void